Update of functions
Stop events on delete Dep/Loc with children

Updated to use server vs dom on data movement

+++ php/deleteDepartmentByID.php
@@ -33,8 +33,36 @@
 
 	}	
 
-	// SQL statement accepts parameters and so is prepared to avoid SQL injection.
-	// $_REQUEST used for development / debugging. Remember to change to $_POST for production
+	$deleteDeptID = $_REQUEST['deleteDeptID'];
+
+	$checkQuery = "SELECT COUNT(firstName) as employees FROM personnel p WHERE p.departmentID = $deleteDeptID";
+
+	$checkResult = $conn->query($checkQuery);
+
+	$data = [];
+
+	while ($row = mysqli_fetch_assoc($checkResult)) {
+
+		array_push($data, $row);
+
+	}
+
+	 $personnel = $data[0]['employees'];
+
+	if ($personnel > 0) {
+
+		$output['status']['code'] = "400";
+		$output['status']['name'] = "executed";
+		$output['status']['description'] = "delete denied";	
+		$output['data'] = [];
+
+		mysqli_close($conn);
+
+		echo json_encode($output); 
+
+		exit;
+	}
+	
 
 	$query = $conn->prepare('DELETE FROM department WHERE id = ?');
